feat(lint) update to eslint-config-jitsi v6
