+++ .eslintrc.js
@@ -1,6 +1,5 @@
 module.exports = {
-    'extends': [
+    extends: [
         '@jitsi/eslint-config'
-    ],
-    'ignorePatterns': [ '*.d.ts' ]
+    ]
 };
