+++ app.js
@@ -10,6 +10,32 @@ import '@matrix-org/olm';
 
 import 'focus-visible';
 
+/*
+* Safari polyfill for createImageBitmap
+* https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope/createImageBitmap
+*
+* Support source image types: Canvas.
+*/
+if (!('createImageBitmap' in window)) {
+    window.createImageBitmap = function(data) {
+        return new Promise((resolve, reject) => {
+            let dataURL;
+
+            if (data instanceof HTMLCanvasElement) {
+                dataURL = data.toDataURL();
+            } else {
+                reject(new Error('createImageBitmap does not handle the provided image source type'));
+            }
+            const img = document.createElement('img');
+
+            img.addEventListener('load', () => {
+                resolve(img);
+            });
+            img.src = dataURL;
+        });
+    };
+}
+
 // We need to setup the jitsi-local-storage as early as possible so that we can start using it.
 // NOTE: If jitsi-local-storage is used before the initial setup is performed this will break the use case when we use
 // the  local storage from the parent page when the localStorage is disabled. Also the setup is relying that
