+++ package.json
@@ -129,13 +129,11 @@
   },
   "devDependencies": {
     "@babel/core": "7.25.9",
-    "@babel/eslint-parser": "7.25.9",
     "@babel/plugin-transform-private-methods": "7.25.9",
     "@babel/preset-env": "7.25.9",
     "@babel/preset-react": "7.25.9",
-    "@jitsi/eslint-config": "5.0.9",
+    "@jitsi/eslint-config": "6.0.1",
     "@react-native/metro-config": "0.75.5",
-    "@stylistic/eslint-plugin": "2.12.1",
     "@types/amplitude-js": "8.16.5",
     "@types/audioworklet": "0.0.29",
     "@types/dom-screen-wake-lock": "1.0.1",
@@ -161,8 +159,6 @@
     "@types/w3c-image-capture": "1.0.6",
     "@types/w3c-web-hid": "1.0.3",
     "@types/zxcvbn": "4.4.1",
-    "@typescript-eslint/eslint-plugin": "8.19.1",
-    "@typescript-eslint/parser": "8.19.1",
     "@wdio/allure-reporter": "9.4.3",
     "@wdio/cli": "9.4.3",
     "@wdio/globals": "9.4.3",
