+++ react/.eslintrc-react-native.js
@@ -1,12 +1,11 @@
 module.exports = {
-    'plugins': [
+    plugins: [
         'react-native'
     ],
-    'rules': {
+    rules: {
         'react-native/no-color-literals': 2,
         'react-native/no-inline-styles': 2,
         'react-native/no-unused-styles': 2,
         'react-native/split-platform-components': 2
     }
 };
-
