+++ react/features/screenshot-capture/ScreenshotCaptureSummary.tsx
@@ -1,5 +1,4 @@
 import 'image-capture';
-import './createImageBitmap';
 
 import { createScreensharingCaptureTakenEvent } from '../analytics/AnalyticsEvents';
 import { sendAnalytics } from '../analytics/functions';
