+++ webpack.config.js
@@ -280,7 +280,7 @@ module.exports = (_env, argv) => {
     };
 
     return [
-        Object.assign({}, config, {
+        { ...config,
             entry: {
                 'app.bundle': './app.js'
             },
@@ -301,10 +301,8 @@ module.exports = (_env, argv) => {
                 })
             ],
 
-            performance: getPerformanceHints(perfHintOptions, 5 * 1024 * 1024)
-
-        }),
-        Object.assign({}, config, {
+            performance: getPerformanceHints(perfHintOptions, 5 * 1024 * 1024) },
+        { ...config,
             entry: {
                 'alwaysontop': './react/features/always-on-top/index.tsx'
             },
@@ -312,9 +310,8 @@ module.exports = (_env, argv) => {
                 ...config.plugins,
                 ...getBundleAnalyzerPlugin(analyzeBundle, 'alwaysontop')
             ],
-            performance: getPerformanceHints(perfHintOptions, 800 * 1024)
-        }),
-        Object.assign({}, config, {
+            performance: getPerformanceHints(perfHintOptions, 800 * 1024) },
+        { ...config,
             entry: {
                 'close3': './static/close3.js'
             },
@@ -322,24 +319,21 @@ module.exports = (_env, argv) => {
                 ...config.plugins,
                 ...getBundleAnalyzerPlugin(analyzeBundle, 'close3')
             ],
-            performance: getPerformanceHints(perfHintOptions, 128 * 1024)
-        }),
+            performance: getPerformanceHints(perfHintOptions, 128 * 1024) },
 
-        Object.assign({}, config, {
+        { ...config,
             entry: {
                 'external_api': './modules/API/external/index.js'
             },
-            output: Object.assign({}, config.output, {
+            output: { ...config.output,
                 library: 'JitsiMeetExternalAPI',
-                libraryTarget: 'umd'
-            }),
+                libraryTarget: 'umd' },
             plugins: [
                 ...config.plugins,
                 ...getBundleAnalyzerPlugin(analyzeBundle, 'external_api')
             ],
-            performance: getPerformanceHints(perfHintOptions, 95 * 1024)
-        }),
-        Object.assign({}, config, {
+            performance: getPerformanceHints(perfHintOptions, 95 * 1024) },
+        { ...config,
             entry: {
                 'face-landmarks-worker': './react/features/face-landmarks/faceLandmarksWorker.ts'
             },
@@ -347,10 +341,8 @@ module.exports = (_env, argv) => {
                 ...config.plugins,
                 ...getBundleAnalyzerPlugin(analyzeBundle, 'face-landmarks-worker')
             ],
-            performance: getPerformanceHints(perfHintOptions, 1024 * 1024 * 2)
-        }),
-        Object.assign({}, config, {
-            /**
+            performance: getPerformanceHints(perfHintOptions, 1024 * 1024 * 2) },
+        { ...config, /**
              * The NoiseSuppressorWorklet is loaded in an audio worklet which doesn't have the same
              * context as a normal window, (e.g. self/window is not defined).
              * While running a production build webpack's boilerplate code doesn't introduce any
@@ -380,10 +372,9 @@ module.exports = (_env, argv) => {
                 ...config.output,
 
                 globalObject: 'AudioWorkletGlobalScope'
-            }
-        }),
+            } },
 
-        Object.assign({}, config, {
+        { ...config,
             entry: {
                 'screenshot-capture-worker': './react/features/screenshot-capture/worker.ts'
             },
@@ -391,7 +382,6 @@ module.exports = (_env, argv) => {
                 ...config.plugins,
                 ...getBundleAnalyzerPlugin(analyzeBundle, 'screenshot-capture-worker')
             ],
-            performance: getPerformanceHints(perfHintOptions, 30 * 1024)
-        })
+            performance: getPerformanceHints(perfHintOptions, 30 * 1024) }
     ];
 };
